--- src/main/kotlin/org/team/b4/cosmicadventures/common/config/SecurityConfig.kt
@@ -1,50 +1,0 @@
-//package org.team.b4.cosmicadventures.common.config
-//
-//import org.springframework.context.annotation.Bean
-//import org.springframework.context.annotation.Configuration
-//import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
-//import org.springframework.security.config.annotation.web.builders.HttpSecurity
-//import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-//import org.springframework.security.web.SecurityFilterChain
-//import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
-//
-//@Configuration
-//@EnableWebSecurity
-//@EnableMethodSecurity
-//class SecurityConfig(
-//    private val jwtAuthenticationFilter: JwtAuthenticationFilter,
-//    private val authenticationEntryPoint: CustomAuthenticationEntryPoint
-//) {
-//
-//    @Bean
-//    fun filterChain(http: HttpSecurity): SecurityFilterChain {
-//        return http
-//            .httpBasic { it.disable() }
-//            .formLogin { it.disable() }
-//            .csrf { it.disable() }
-//            .authorizeHttpRequests {
-//                it.requestMatchers(
-//                    "/login",
-//                    "/signup",
-//                    "/swagger-ui/**",
-//                    "/v3/api-docs/**",
-//                ).permitAll()
-//                    .requestMatchers("/admin/**").hasRole("ADMIN")
-//                    // 위 URI를 제외하곤 모두 인증이 되어야 함.
-//                    .anyRequest().authenticated()
-//            }
-//            .logout {
-//                it.logoutUrl("/logout")
-//                    .logoutSuccessUrl("/swagger-ui/index.html")
-//                    .deleteCookies("Set-Cookie")
-//                    .invalidateHttpSession(true)
-//            }
-//            // 기존 UsernamePasswordAuthenticationFilter 가 존재하던 자리에 JwtAuthenticationFilter 적용
-//            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)//
-//            .exceptionHandling {
-//                it.authenticationEntryPoint(authenticationEntryPoint)
-//            }
-//            .build()
-//    }
-//
-//}
