--- src/main/kotlin/org/team/b4/cosmicadventures/common/config/SwaggerConfig.kt
@@ -1,37 +1,0 @@
-package org.team.b4.cosmicadventures.common.config
-
-import io.swagger.v3.oas.models.Components
-import io.swagger.v3.oas.models.OpenAPI
-import io.swagger.v3.oas.models.info.Info
-import io.swagger.v3.oas.models.security.SecurityRequirement
-import io.swagger.v3.oas.models.security.SecurityScheme
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-@Configuration
-class SwaggerConfig {
-
-    @Bean
-    fun openAPI(): OpenAPI {
-        return OpenAPI()
-            .addSecurityItem(
-                SecurityRequirement().addList("Bearer Authentication")
-            )
-            .components(
-                Components().addSecuritySchemes(
-                    "Bearer Authentication",
-                    SecurityScheme()
-                        .type(SecurityScheme.Type.HTTP)
-                        .scheme("Bearer")
-                        .bearerFormat("JWT")
-                        .`in`(SecurityScheme.In.HEADER)
-                        .name("Authorization")
-                )
-            )
-            .info(
-                Info()
-                    .title("Space API")
-                    .description("Space API schema")
-                    .version("1.0.0")
-            )
-    }
-}
