+++ src/main/kotlin/org/team/b4/cosmicadventures/global/config/SwaggerConfig.kt
@@ -1,4 +1,4 @@
-package org.team.b4.cosmicadventures.global.swagger
+package org.team.b4.cosmicadventures.global.config
 
 import io.swagger.v3.oas.models.Components
 import io.swagger.v3.oas.models.OpenAPI
