--- src/main/kotlin/org/team/b4/cosmicadventures/specialDay/dto/TestRequest.kt
@@ -1,10 +1,0 @@
-package org.team.b4.cosmicadventures.specialDay.dto
-
-
-data class TestRequest (
-    val serviceKey: String,
-    val pageNo: Int,
-    val numOfRows: Int,
-    val solYear: Int,
-    val solMonth: Int
-)
