+++ src/main/resources/application.yml
@@ -1,4 +1,35 @@
+spring:
+  # DB Settings
+  datasource:
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:~/space
+    username: hyeonseok
+    password:
+  #  datasource:
+  cache:
+    type: redis
+  data:
+    redis:
+      host: localhost  # Redis 호스트 주소
+      port: 6379       # Redis 포트
+      database: 0      # Redis 데이터베이스 번호
+      lettuce:
+        pool:
+          max-active: 8  # 최대 활성 커넥션 수
+          max-idle: 8    # 최대 유휴 커넥션 수
+          min-idle: 2    # 최소 유휴 커넥션 수
+          max-wait: -1   # 최대 대기 시간 (-1은 무제한)
+      password:
+  # Hibernate
+  jpa:
+    hibernate:
+      ddl-auto: update
+auth:
+  jwt:
+    issuer: team.sparta.com
+    secret: 52DANSJDNASJnjdgsnjFDNJSDFSDFDSF
+    accessTokenExpirationHour: 168
+# Log
 logging:
   level:
-    org.springframework.web.reactive.function.client.ExchangeFunctions: DEBUG
     org.hibernate.sql: debug
